adding start of work to clean expired jobs
this will be run as a nightly task, and will push to master
after checking over the list of jobs and only checking those that are
expired, and of that set, it it does not return a successful response
we remove it from the list. I am opening this with a trigger on PR so
that I can test, and also am not allowing the push to master to happen.
When the script seems to run okay I will remove the pull request trigger
and then the exit code before push, and we will test fully for the first
overnight run.

Signed-off-by: vsoch <vsochat@stanford.edu>