removing map endpoint in favor of usrse-map.
The data and page still exists, but the testing has been
disabled to speed up the CI. This is being done in favor
of completely deleting the includes and data files, in
case we want to create another kind of map page. It does
no damage to the site or performance cuts to keep the page
present (but hidden).

Signed-off-by: vsoch <vsochat@stanford.edu>