adding first US-RSE history post!
The idea I (@vsoch) came up with while preparing a talk about the history of US-RSE, I mentioned
to Dan and he wrote an initial skeleton for his blog, but there was misunderstanding and I explained
we wanted a community post, so then I offered to flush it out into a community document and invited others,
adding much more detail, history, and stories. Now we have this lovely post and I am very proud of our community!
Although it was my (@vsoch) idea and I did most of the refactoring I am fine to have dan as first author on
the final thing since he wrote the original brief description.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>