adding support for last-modified (#759)
* adding support for last-modified

this includes a new "include" that will allow us to set a boolean, set_last_modified to true on any
page to automatically add a last modified date to the bottom left, courtesy of the GitHub commits api. Given that
we want to customize either 1. the div id, 2. the repository used (and in this case) 3. the path, those are
also exposed as variables to customize (e.g., current set for the code of conduct and governance). Including these
changes to templates I have also added the last modified date to be included on most about pages and working group
pages, and double checked that the dates are accurate. Thanks Peter for the great idea!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* Removing last modified in governanace

Co-authored-by: vsoch <vsoch@users.noreply.github.com>