Closes #1068 - Use current time zone when creating calendar entries
The "Add to Calendar" button (per issue #1068) was creating entries anchored in "America/New_York" which meant it was shifting the time to the other time zone for non-Eastern TZ folks.  This uses the current time zone for the user.

Credit to @exoticDFT for the actual code fix to use