Fix issue with social buttons.
The `site.url` variable needs to be the site domain.
The absence of a value causes URLs using `site.url` to not
contain this, which then prevents social sites from linking
to the site.

This change fixes the `site.url` variable and maintains
the circleci build behavior to maintain preview of assets.

Closes #1201