Add event for Oct2023 Funder Talk Series (Schmidt Futures)
In addition to the event entry and an update to the Outreach WG page, this adds a sub-folder for images (headshots).  In the past we have linked to ones on the
internet, but for this event received headshots that could not be found/linked and have to be hosted in our site.