Create jobs-form-exporter.gs
put a copy of the jobs form exporter into the repo for posterity
    
This Google Apps Script currently runs in my personal Google
account (and only sends email to me at the moment).  Placing this into the repo allows someone else from US-RSE to recreate it in the event that my Google account goes away.